feat: Moved transfer code to TransferSession component and implemented at ContactInfo
